pyproject.toml
[tool.poetry]
name = "fastapi-auth"
version = "0.1.0"
description = ""
authors = ["ryanvincere <ryan@continuum.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.1"
fastapi = "^0.110.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
